[caffe2][nomnigraph] Generic subgraph replacement (#6368)
* Nomnigraph cutting

* Comments

* Use caffe2::to_string

* More to_string

+++ caffe2/core/nomnigraph/include/nomnigraph/Converters/Dot.h
@@ -54,7 +54,7 @@ public:
       output << "subgraph cluster" << i << " {\n";
       output << "style=dotted;\n";
       for (const auto &node : subgraph->getNodes()) {
-        output << static_cast<uint64_t>(node);
+        output << node;
         output << ";\n";
       }
       output << "}\n";
